+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 
 ##### Enhancements
 
-* Return `SWIFT_NAME` when generating Objective-C docs.
+* Return `SWIFT_NAME` when generating Objective-C docs.  
   [Ibrahim Ulukaya](https://github.com/ulukaya)
 
 ##### Bug Fixes