+++ CHANGELOG.md
@@ -10,7 +10,10 @@
 
 ##### Bug Fixes
 
-* None.
+* Fix Swift Package Manager warnings when using SourceKitten targets as
+  dependencies.  
+  [JP Simard](https://github.com/jpsim)
+  [#478](https://github.com/jpsim/SourceKitten/issues/478)
 
 ## 0.19.0